requirements.txt
numpy
streamlit
matplotlib
opencv-python-headless
streamlit-drawable-canvas==0.8.0
scikit-learn
google-cloud-firestore
firebase_admin
scipy
PyQt6
torch
tqdm
redis